Split vpts module in submodules
